Edit def on() 2nd time
Delete 3 print functions. Read used.csv( from the path to retrieve the file) I and squeezes the row no. of the series. If the password generated was existing on used.csv, it will return to the initial of on() and generate password again. In contrast, the password generated will be embedded in the used.csv without overwrite existing password